server/requirements.txt
# ── LangGraph / LangChain ────────────────────────────────────────────────────
langgraph>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0           # ChatOpenAI — used for GitHub Models endpoint

# ── Blockchain / crypto ──────────────────────────────────────────────────────
web3>=6.0.0
eth-account>=0.11.0
ecdsa>=0.18.0

# ── x402 payment protocol ───────────────────────────────────────────────────
x402>=0.1.0

# ── HTTP / async ─────────────────────────────────────────────────────────────
httpx>=0.27.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── Database / ORM ───────────────────────────────────────────────────────────
prisma>=0.13.0

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0

# ── Required dependencies ─────────────────────────────────────────────────────
fastapi
uvicorn
langchain
langgraph
python-dotenv
httpx